+++ test/properties.js
@@ -54,6 +54,7 @@ describe('58. properties.js', function() {
       defaultValues.outFormat        = oracledb.outFormat;
       defaultValues.lobPrefetchSize  = oracledb.lobPrefetchSize;
       defaultValues.queueTimeout     = oracledb.queueTimeout;
+      defaultValues.queueMax         = oracledb.queueMax;
       defaultValues.stmtCacheSize    = oracledb.stmtCacheSize;
       defaultValues.poolPingInterval = oracledb.poolPingInterval;
       defaultValues.fetchAsBuffer    = oracledb.fetchAsBuffer;
@@ -75,6 +76,7 @@ describe('58. properties.js', function() {
       oracledb.outFormat        = defaultValues.outFormat;
       oracledb.lobPrefetchSize  = defaultValues.lobPrefetchSize;
       oracledb.queueTimeout     = defaultValues.queueTimeout;
+      oracledb.queueMax         = defaultValues.queueMax;
       oracledb.stmtCacheSize    = defaultValues.stmtCacheSize;
       oracledb.poolPingInterval = defaultValues.poolPingInterval;
       oracledb.fetchAsBuffer    = defaultValues.fetchAsBuffer;
@@ -210,7 +212,15 @@ describe('58. properties.js', function() {
       should.notEqual(oracledb.queueTimeout, defaultValues.queueTimeout);
     });
 
-    it('58.1.16 stmtCacheSize', function() {
+    it('58.1.16 queueMax', function() {
+      var t = oracledb.queueMax;
+      oracledb.queueMax = t + 1000;
+
+      should.equal(t, defaultValues.queueMax);
+      should.notEqual(oracledb.queueMax, defaultValues.queueMax);
+    });
+
+    it('58.1.17 stmtCacheSize', function() {
       var t = oracledb.stmtCacheSize;
       oracledb.stmtCacheSize = t + 5;
 
@@ -218,7 +228,7 @@ describe('58. properties.js', function() {
       should.notEqual(oracledb.stmtCacheSize, defaultValues.stmtCacheSize);
     });
 
-    it('58.1.17 poolPingInterval', function() {
+    it('58.1.18 poolPingInterval', function() {
       var t = oracledb.poolPingInterval;
       oracledb.poolPingInterval = t + 100;
 
@@ -226,7 +236,7 @@ describe('58. properties.js', function() {
       should.notEqual(oracledb.poolPingInterval, defaultValues.poolPingInterval);
     });
 
-    it('58.1.18 fetchAsBuffer', function() {
+    it('58.1.19 fetchAsBuffer', function() {
       var t = oracledb.fetchAsBuffer;
       oracledb.fetchAsBuffer = [ oracledb.BLOB ];
 
@@ -234,7 +244,7 @@ describe('58. properties.js', function() {
       (oracledb.fetchAsBuffer).should.not.eql(defaultValues.fetchAsBuffer);
     });
 
-    it('58.1.19 Negative - connectionClass ', function() {
+    it('58.1.20 Negative - connectionClass ', function() {
       should.throws(
         function() {
           oracledb.connectionClass = NaN;
@@ -243,7 +253,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.20 Negative - autoCommit', function() {
+    it('58.1.21 Negative - autoCommit', function() {
       should.throws(
         function() {
           oracledb.autoCommit = 2017;
@@ -252,7 +262,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.21 Negative - outFormat', function() {
+    it('58.1.22 Negative - outFormat', function() {
       should.throws(
         function() {
           oracledb.outFormat = 'abc';
@@ -261,7 +271,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.22 Negative - externalAuth', function() {
+    it('58.1.23 Negative - externalAuth', function() {
       should.throws(
         function() {
           oracledb.externalAuth = 2017;
@@ -270,7 +280,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.23 versionString (read-only)', function() {
+    it('58.1.24 versionString (read-only)', function() {
       var t = oracledb.versionString;
       t.should.be.a.String();
 
@@ -282,7 +292,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.24 versionSuffix (read-only)', function() {
+    it('58.1.25 versionSuffix (read-only)', function() {
       var t = oracledb.versionSuffix;
 
       if (t) // it could be a String, or undefined
@@ -296,7 +306,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.25 oracleClientVersionString (read-only)', function () {
+    it('58.1.26 oracleClientVersionString (read-only)', function () {
       var t = oracledb.oracleClientVersionString;
       t.should.be.a.String();
 
@@ -308,7 +318,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.26 edition', function() {
+    it('58.1.27 edition', function() {
       var t = oracledb.edition;
       oracledb.edition = 'foobar';
       var e = oracledb.edition;
@@ -317,7 +327,7 @@ describe('58. properties.js', function() {
       should.strictEqual(e, 'foobar');
     });
 
-    it('58.1.27 Negative - edition', function() {
+    it('58.1.28 Negative - edition', function() {
       should.throws(
         function() {
           oracledb.edition = 123;
@@ -326,7 +336,7 @@ describe('58. properties.js', function() {
       );
     });
 
-    it('58.1.28 events', function() {
+    it('58.1.29 events', function() {
       var t = oracledb.events;
       oracledb.events = true;
 
@@ -334,7 +344,7 @@ describe('58. properties.js', function() {
       should.strictEqual(oracledb.events, true);
     });
 
-    it('58.1.29 Negative - events', function() {
+    it('58.1.30 Negative - events', function() {
       should.throws(
         function() {
           oracledb.events = 'hello';
@@ -475,6 +485,18 @@ describe('58. properties.js', function() {
           pool.poolPingInterval = t + 100;
         },
         "TypeError: Cannot assign to read only property 'poolPingInterval' of object '#<Pool>"
+      );
+    });
+
+    it('58.2.10 queueMax', function() {
+      var t = pool.queueMax;
+      t.should.be.a.Number();
+
+      should.throws(
+        function() {
+          pool.queueMax = t + 1000;
+        },
+        "TypeError: Cannot assign to read only property 'queueMax' of object '#<Pool>"
       );
     });